CB-6212 iOS: fix warnings compiled under arm64 64-bit
-fix one more warning

(cherry picked from commit db9fa24cf8ba165aabae0460a9cbeb1dd1d67e62)

+++ src/ios/CDVCapture.m
@@ -815,7 +815,7 @@
 
     if (!self.pluginResult) {
         // return error
-        self.pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageToErrorObject:self.errorCode];
+        self.pluginResult = [CDVPluginResult resultWithStatus:CDVCommandStatus_ERROR messageToErrorObject:(int)self.errorCode];
     }
 
     self.avRecorder = nil;
